refactor: Move tableofcontents data to index.ts
